allow chaining on setters (#57532)
this allows calling code to chain allowing for more fluent code.

this code was last touched 10 years ago, which is probably why it isn't chained already, since it seems like such a common scenario for chaining.

these methods are not used a ton in the codebase, but this would allow us to clean up the tests a little.